==> lapse-analysis/01_Transtion_Matrix.ipynb <==
import os
import io
import pandas as pd
import numpy as np
import glob
import matplotlib.pyplot as plt
import seaborn as sns; sns.set()

import datetime

#dask
import dask.dataframe as dd

import warnings
warnings.filterwarnings("ignore")

### Path

path = r"\\VNCRP1002\\ud03$\\X266754\\My Documents\\Lapse\\Data\\"

### PFMC

filenames = glob.glob(path + "\*pfmc*.csv")
print(filenames)

#list of all df
dfs = []
for f in filenames: 
    print(f)
    d = pd.read_csv(f, encoding="ISO-8859-1") 
    dfs.append(d) 
pfmc = pd.concat(dfs, ignore_index=True)
del dfs

# remove duplicates
pfmc = pfmc.drop_duplicates()

# create current and previous months for pfmc
# pfmc cur_month,  pfmc_prev_month
pfmc['val_dt']= pfmc['val_dt'].astype('datetime64[ns]')
pfmc['cur_month'] = pfmc['val_dt'].astype('datetime64[ns]').map(lambda x: x.strftime('%Y-%m'))
pfmc['prev_month'] =  ((pfmc['val_dt'] - pd.DateOffset(months=1)).dt.to_period("M")).astype(str)

pd.set_option('display.max_columns', None)
#pfmc.head()
pfmc[pfmc['agmt_pkge_id']=='V30059210000'][['agmt_pkge_id','prev_month', 'cur_month']]

### Events_Monthly

filenames = glob.glob(path + "\*events_monthly*.csv")
print(filenames)

#list of all df
dfs = []
for f in filenames: 
    print(f)
    d = pd.read_csv(f, encoding="ISO-8859-1") 
    dfs.append(d) 
events = pd.concat(dfs, ignore_index=True)
del dfs

# remove duplicates
events = events.drop_duplicates()

# create current and previous months for events 
events['prcs_dt']= events['prcs_dt'].astype('datetime64[ns]')
events['evnt_cur_month'] = events['prcs_dt'].map(lambda x: x.strftime('%Y-%m'))
events['evnt_prev_month'] =  ((events['prcs_dt'] - pd.DateOffset(months=1)).dt.to_period("M")).astype(str)

# example
pd.set_option('display.max_columns', None)
events[events['agmt_pkge_id']=='V30059210000'][['agmt_pkge_id','evnt_desc','evnt_prev_month', 'evnt_cur_month']]

### Join PFMC Inf and Events to get Lapse policiies

# Merge pfmv and events
df = pfmc.merge(events[['agmt_pkge_id','evnt_desc','evnt_prev_month', 'evnt_cur_month']], 
              how='left', 
              left_on=['agmt_pkge_id', 'cur_month'], 
              right_on=['agmt_pkge_id', 'evnt_prev_month'], 
              suffixes=('', '_remove'))
df.drop([i for i in df.columns if 'remove' in i],axis=1, inplace=True)

df.shape

# check
df[df['agmt_pkge_id']=='V30059210000'][['agmt_pkge_id','evnt_desc',
                                        'prev_month', 'cur_month',
                                        'evnt_prev_month', 'evnt_cur_month']]

#del pfmc, events

### Lapse and Surrender Indicators (reinstatement, to be done)

# forward fill indicators
df['evnt_desc']= df.groupby('agmt_pkge_id')['evnt_desc'].ffill()

# indicator
df['lapse_ind'] = (np.where(df['evnt_desc'].str.contains('Lapse', na=False),1, 0)).astype(int)
df['surrender_ind'] = (np.where(df['evnt_desc'].str.contains('Surrender',na=False),1, 0)).astype(int)

## create lag indicators
df['lapse_ind']= df['lapse_ind'].fillna(0)
df['lapse_ind_lag1'] = df.groupby('agmt_pkge_id')['lapse_ind'].shift(1)
df['lapse_ind_lag1']=df['lapse_ind_lag1'].fillna(0)

df['surrender_ind']= df['surrender_ind'].fillna(0)
df['surrender_ind_lag1'] = df.groupby('agmt_pkge_id')['surrender_ind'].shift(1)
df['surrender_ind_lag1']=df['surrender_ind_lag1'].fillna(0)

df['cur_status']= np.where(df['lapse_ind']==1, 'lapse', 
                           (np.where(df['surrender_ind']==1, 'surrender', "current")))

df['prior_status']= np.where(df['lapse_ind_lag1']==1, 'lapse', 
                           (np.where(df['surrender_ind_lag1']==1, 'surrender', "current")))

# check
df[df['agmt_pkge_id']=='965390970000'][['agmt_pkge_id','evnt_desc','lapse_ind', 'surrender_ind',
                                        'prior_status', 'cur_status',
                                        'prev_month', 'cur_month',
                                        'evnt_prev_month', 'evnt_cur_month']]

dgroup = df.groupby(["cur_month", "lapse_ind"], as_index=True)['lapse_ind'].count()
dgroup

dgroup = df.groupby(["cur_month", "surrender_ind"], as_index=True)['agmt_pkge_id'].count()
dgroup

### Transition Matrix

sample = df[df['cur_month']=='2022-08']
tmatrix = pd.crosstab(index=sample['prior_status'], columns=[sample['cur_status']]).reset_index()
tmatrix

pd.crosstab(index=sample['prior_status'], columns=[sample['cur_status']],normalize='index').round(4)*100

#### check

df[(df['prior_status']=='current')&(df['cur_status']=='surrender')].head()

#### export to parquet

df=df[df['cur_month']<='2022-08']
df.to_parquet('pfmc_events_apr_aug.parquet')
