==> lapse-analysis/1_PFMC_Processing.ipynb <==
!pip install awswrangler --q

import os
import io
import pandas as pd
import numpy as np
import glob
import matplotlib.pyplot as plt
import seaborn as sns; sns.set()

import datetime

import awswrangler as wr
#import dask.dataframe as dd

import warnings
warnings.filterwarnings("ignore")

### local processing 

path = "/root/Lapse/data/"

#### pfmc inf

filenames = glob.glob(path + "*pfmc*.csv")
print(filenames)

#list of all df
dfs = []
for f in filenames: 
    print(f)
    d = pd.read_csv(f, encoding="ISO-8859-1") 
    dfs.append(d) 
pfmc = pd.concat(dfs, ignore_index=True)
del dfs
print(pfmc.shape)

pfmc['val_dt']= pfmc['val_dt'].astype('datetime64[ns]')

# remove NAT time 
pfmc = pfmc.dropna(subset=['val_dt'])
pfmc['pfmc_cur_month'] = pfmc['val_dt'].map(lambda x: x.strftime('%Y-%m'))

pfmc.groupby('pfmc_cur_month')['agmt_pkge_id'].count()

#### pfmc events

filenames = glob.glob(path + "*events_monthly*.csv")
print(filenames)

#list of all df
dfs = []
for f in filenames: 
    print(f)
    d = pd.read_csv(f, encoding="ISO-8859-1") 
    dfs.append(d) 
events = pd.concat(dfs, ignore_index=True)
del dfs

#import dateutil.relativedelta #evnt_dt or prcs_dt?
events['prcs_dt']= events['prcs_dt'].astype('datetime64[ns]')
events['evnt_cur_month'] = events['prcs_dt'].map(lambda x: x.strftime('%Y-%m'))
events['evnt_3_month_ago'] =  ((events['prcs_dt'] - pd.DateOffset(months=3)).dt.to_period("M")).astype(str)

# indicator
events['lapse_ind'] = (np.where(events['evnt_desc'].str.contains('Lapse', na=False),1, 0)).astype(int)
events['surrender_ind'] = (np.where(events['evnt_desc'].str.contains('Surrender',na=False),1, 0)).astype(int)
events['reinstate_ind'] = (np.where(events['evnt_desc'].str.contains('True Reinstatement',na=False),1, 0)).astype(int)

### join 

# Merge pfmv and events
df = pfmc.merge(events[['agmt_pkge_id','evnt_desc','lapse_ind', 'surrender_ind', 'reinstate_ind','evnt_3_month_ago', 
                        'evnt_cur_month']], 
              how='left', 
              left_on=['agmt_pkge_id', 'pfmc_cur_month'], 
              right_on=['agmt_pkge_id', 'evnt_3_month_ago'], 
              suffixes=('', '_remove'))

df.drop([i for i in df.columns if 'remove' in i],axis=1, inplace=True)

print(df.shape)

# this causes mixed type
del df["iss_rmo"]
del df['limra_lob_desc']

### write back as parquet

# import pyarrow
# import pyarrow.parquet as pq

# table = pyarrow.Table.from_pandas(df)

%%time
import awswrangler as wr

wr.s3.to_parquet(
    df=df,
    path='s3://cdo-ililapse-364524684987-bucket/x266754/lapse/pfmc_jan_jun.parquet'
    #dataset=True
)

### Aws Set-up

session = boto3.Session()
s3 = session.client('s3')

### Test: Read parquet from S3 bucket location

bucket_name = "cdo-ililapse-364524684987-bucket"
file_path = "x266754/lapse/"
#s3_path = "s3://"+bucket_name+file_path
response = s3.list_objects(Bucket = bucket_name, Prefix = file_path + 'pfmc')

%%time

df = []
for i in range(len(response['Contents'])):
    key = response['Contents'][i]['Key']
    obj = s3.get_object(Bucket = bucket_name, Key = key)
    df_temp = pd.read_parquet(io.BytesIO(obj['Body'].read()))
    df.append(df_temp)
    print(i,' = ',response['Contents'][i]['Key'])
df = pd.concat(df, ignore_index = True)

df.head()

df.shape
