==> lapse-analysis/2_ECIW_Processing.ipynb <==
!pip install awswrangler --q

import os
import io
import pandas as pd
import numpy as np
import glob
import matplotlib.pyplot as plt
import seaborn as sns; sns.set()

import datetime

import awswrangler as wr
#import dask.dataframe as dd

import warnings
warnings.filterwarnings("ignore")

#### eciw household 

path = "/root/Lapse/data/"

filenames = glob.glob(path + "*household*.csv")
print(filenames)

#list of all df
dfs = []
for f in filenames: 
    print(f)
    d = pd.read_csv(f, encoding="ISO-8859-1") 
    dfs.append(d) 
eciw_hh = pd.concat(dfs, ignore_index=True)
del dfs
print(eciw_hh.shape)

eciw_hh['as_of_hist_dt']=eciw_hh['as_of_hist_dt'].astype('datetime64[ns]')
eciw_hh['hhld_acq_dt']=eciw_hh['hhld_acq_dt'].astype('datetime64[ns]')
eciw_hh['most_rcnt_pur_dt']=eciw_hh['most_rcnt_pur_dt'].astype('datetime64[ns]')

eciw_hh = eciw_hh[['household_id', 
               #dates
               'as_of_hist_dt', 'hhld_acq_dt', 'most_rcnt_pur_dt',
               # codes
               'incm_rng_cd', 'prim_occup_cd','prmr_prsnc_chld_ind', 'net_wrth_gold_cd',
             'hhld_qdrnt_cd',
               #asset mix
              'wc_tot_asset_amt','wc_asset_mix_anty_amt','wc_asset_mix_bond_amt',
               'wc_asset_mix_dep_amt','wc_asset_mix_mf_amt','wc_asset_mix_oth_asset_amt',
               'wc_asset_mix_stk_amt']]

#### eciw subject 

filenames = glob.glob(path + "*subject*.csv")
print(filenames)

#list of all df
dfs = []
for f in filenames: 
    print(f)
    d = pd.read_csv(f, encoding="ISO-8859-1") 
    dfs.append(d) 
eciw_sb = pd.concat(dfs, ignore_index=True)
del dfs
print(eciw_sb.shape)

eciw_sb['as_of_hist_dt']=eciw_sb['as_of_hist_dt'].astype('datetime64[ns]')
eciw_sb['subj_acq_dt']=eciw_sb['subj_acq_dt'].astype('datetime64[ns]')

eciw_sb = eciw_sb[['household_id', 'subject_id', 'agmt_pkge_id','policy_id',
               'as_of_hist_dt', 'subj_acq_dt',
               'mrtl_stat_cd', 'gndr_cd', 'subj_state_cd', 'subj_act_ind',
               'pref_lang_cd','deceased_ind', 'rtl_dstrb_orphn_cd',
               'ixi_buoyancy', 'ixi_fin_cohort_cd', 
               'subj_aum_amt' ]]

#### join 

df =eciw_sb.merge(eciw_hh, how='left', 
                left_on=['household_id', 'as_of_hist_dt'], 
                right_on = ['household_id', 'as_of_hist_dt'],
                suffixes=('', '_remove'))
df.drop([i for i in df.columns if 'remove' in i],axis=1, inplace=True)

df = df.sort_values(['household_id', 'as_of_hist_dt'])

df = df.drop_duplicates()

print(df.shape)

#### export

%%time
import awswrangler as wr

wr.s3.to_parquet(
    df=df,
    path='s3://cdo-ililapse-364524684987-bucket/x266754/lapse/eciw_jan_jun.parquet'
    #dataset=True
)
