==> lapse-analysis/3_Clarify_Processing.ipynb <==
!pip install awswrangler --q

import sagemaker
import io
import os

from sagemaker.sklearn.processing import SKLearnProcessor
from sagemaker.network import NetworkConfig
from sagemaker.processing import ProcessingInput, ProcessingOutput

import pandas as pd
import numpy as np
import boto3
import warnings
warnings.filterwarnings("ignore")

### set-up

session = boto3.Session()
s3 = session.client('s3')

### bucket location

bucket_name = "cdo-ililapse-364524684987-bucket"
file_path = "nfrm_data/clarify/2022/"
#s3_path = "s3://"+bucket_name+file_path
response = s3.list_objects(Bucket = bucket_name, Prefix = file_path + 'clarify_report_sr_notes')

%%time

df = []
for i in range(len(response['Contents'])):
    key = response['Contents'][i]['Key']
    obj = s3.get_object(Bucket = bucket_name, Key = key)
    df_temp = pd.read_csv(io.BytesIO(obj['Body'].read()), encoding="ISO-8859-1",header=None, skiprows=1, na_values=" ")
    df.append(df_temp)
    print(i,' = ',response['Contents'][i]['Key'])
df = pd.concat(df, ignore_index = True)

df.columns =[
'POLICY_NBR',
'CASE_ID',
'CASE_TITLE',
'CASE_CREATE_DATE',
'CLIENT_ID',
'SR_ID',
'SR_TITLE',
'SR_TYPE_CATG',
'SR_TYPE_DESC',
'SUB_TYPE',
'SR_CREATE_DT',
'SR_CLOSE_DT',
'ACTION_TYPE',
'SR_NOTES'
]

df=df.sort_values(['POLICY_NBR', 'SR_CREATE_DT'])

pd.set_option('display.max_colwidth', 1000)
sample = df[df['POLICY_NBR']=='V3077010']
sample.to_csv("clarify_sample.csv")

### processing

df_clarify = df[['POLICY_NBR', 'SR_CREATE_DT', 'SR_CLOSE_DT','SR_TYPE_CATG']]

# remove whitespaces
for col in df_clarify.columns:
    try:
        df_clarify[col] = df_clarify[col].str.strip()
    except AttributeError:
        pass

df_clarify["SR_TYPE_CATG"]= df_clarify["SR_TYPE_CATG"].str.replace(" ","")

df_clarify['SR_CREATE_DT']= pd.to_datetime(df_clarify['SR_CREATE_DT'], format="%Y-%m-%d-%H.%M.%S.%f").dt.strftime('%m/%d/%y')
df_clarify['SR_CREATE_DT']=pd.to_datetime(df_clarify['SR_CREATE_DT'])

df_clarify['SR_CLOSE_DT']= pd.to_datetime(df_clarify['SR_CLOSE_DT'], format="%Y-%m-%d-%H.%M.%S.%f").dt.strftime('%m/%d/%y')
df_clarify['SR_CLOSE_DT']=pd.to_datetime(df_clarify['SR_CLOSE_DT'])

#### max time to process

min_create_dt= pd.DataFrame(df_clarify.groupby(['POLICY_NBR','SR_TYPE_CATG'])['SR_CREATE_DT'].min()).reset_index()
max_close_dt= pd.DataFrame(df_clarify.groupby(['POLICY_NBR','SR_TYPE_CATG'])['SR_CLOSE_DT'].max()).reset_index()
min_max_df = min_create_dt.merge(max_close_dt, how= 'inner', on =['POLICY_NBR','SR_TYPE_CATG'])
min_max_df['sr_create_month'] = min_max_df['SR_CREATE_DT'].dt.to_period('M')

min_max_df['closed'] = np.where(min_max_df['SR_CLOSE_DT'].isnull() , 0, 1)
min_max_df['time_to_close']= np.round((min_max_df['SR_CLOSE_DT'] - min_max_df['SR_CREATE_DT'])/np.timedelta64(1, 'D'), 0)

min_max_df['closed'] = np.where(min_max_df['SR_CLOSE_DT'].isnull() , 0, 1)
min_max_df['time_to_close']= np.round((min_max_df['SR_CLOSE_DT'] - min_max_df['SR_CREATE_DT'])/np.timedelta64(1, 'D'), 0)

#### Number of cases

df_count = pd.DataFrame(min_max_df.groupby(['POLICY_NBR','sr_create_month'])['SR_TYPE_CATG'].count()).reset_index()
df_count.columns = ['POLICY_NBR', 'sr_create_month', 'num_sr_catg']

#### Longest taking SR Category

# service category that took the longest in each month
df_sr_time = pd.DataFrame(min_max_df.groupby(['POLICY_NBR','sr_create_month'])['time_to_close'].max()).reset_index()
df_sr_time.columns = ['POLICY_NBR', 'sr_create_month', 'max_sr_time']

df_out = df_count.merge(df_sr_time, how= 'inner', on =['POLICY_NBR','sr_create_month'])

del df_count, df_sr_time

df_out = df_out.merge(min_max_df, how= 'inner', 
                      left_on =['POLICY_NBR','sr_create_month','max_sr_time'],
                      right_on =['POLICY_NBR','sr_create_month','time_to_close']
                     )

df_out[df_out['POLICY_NBR']=='V3077010']

#### export to S3

# %%time
# df_out =pd.read_csv("clarify_sample.csv")


# import awswrangler as wr
# wr.s3.to_csv(
#     df=df_out,
#     path='s3://cdo-ililapse-364524684987-bucket/x266754/temp/clarify_sample.csv'
# )

%%time
import awswrangler as wr

wr.s3.to_parquet(
    df=df_out,
    path='s3://cdo-ililapse-364524684987-bucket/x266754/lapse/clarify_jan_jun.parquet'
    #dataset=True
)
