==> lapse-analysis/4_Data_Curation.ipynb <==
!pip install awswrangler --q

import boto3
import io
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns; sns.set()

import awswrangler as wr

import warnings
warnings.filterwarnings("ignore")

plt.style.use('seaborn-colorblind')

### set-up

session = boto3.Session()
s3 = session.client('s3')

bucket_name = "cdo-ililapse-364524684987-bucket"
file_path = "x266754/lapse/"

### pfmc

file_name = "x266754/lapse/pfmc_jan_jun.parquet"

%%time
obj = s3.get_object(Bucket = bucket_name, Key = file_name)
pfmc = pd.read_parquet(io.BytesIO(obj['Body'].read())) 

pfmc[pfmc['agmt_pkge_id']=='V26808080000'][['policy_id', 'pfmc_cur_month']]

### eciw

file_name = "x266754/lapse/eciw_jan_jun.parquet"

%%time
obj = s3.get_object(Bucket = bucket_name, Key = file_name)
eciw = pd.read_parquet(io.BytesIO(obj['Body'].read())) 

eciw['eciw_cur_month'] = eciw['as_of_hist_dt'].astype('datetime64[ns]').map(lambda x: x.strftime('%Y-%m'))
eciw['eciw_prev_month'] =  ((eciw['as_of_hist_dt'] - pd.DateOffset(months=1)).dt.to_period("M")).astype(str)

eciw[eciw['agmt_pkge_id']=='V26808080000'][['agmt_pkge_id', 'policy_id', 'eciw_cur_month']]

### merge pfmc and eciw 

%%time
df=pfmc.merge(eciw, how='left', 
              left_on=['agmt_pkge_id','pfmc_cur_month'], 
              right_on = ['agmt_pkge_id','eciw_cur_month'],
              suffixes=('', '_remove'))
df.drop([i for i in df.columns if 'remove' in i],axis=1, inplace=True)

df[df['agmt_pkge_id']=='V26808080000'][['agmt_pkge_id', 'policy_id', 'pfmc_cur_month', 'eciw_cur_month']]

del pfmc, eciw

### clarify

file_name = "x266754/lapse/clarify_jan_jun.parquet"

%%time
obj = s3.get_object(Bucket = bucket_name, Key = file_name)
clarify = pd.read_parquet(io.BytesIO(obj['Body'].read())) 

clarify['clarify_cur_month'] = clarify['SR_CREATE_DT'].astype('datetime64[ns]').map(lambda x: x.strftime('%Y-%m'))

clarify['SR_TYPE_CATG'] = np.where(clarify['SR_TYPE_CATG'].isnull(), 'None', clarify['SR_TYPE_CATG'])

clarify.head()

### merge with clarify

%%time
df=df.merge(clarify, how='left', 
              left_on=['policy_id','pfmc_cur_month'], 
              right_on = ['POLICY_NBR','clarify_cur_month'],
              suffixes=('', '_remove'))
df.drop([i for i in df.columns if 'remove' in i],axis=1, inplace=True)

del clarify

#del df
print(df.shape)

pd.set_option('max_columns', None)
df.head()

### Export back

%%time
import awswrangler as wr

wr.s3.to_parquet(
    df=df2,
    path='s3://cdo-ililapse-364524684987-bucket/x266754/lapse/lapse_curated_jan_jun.parquet'
    #dataset=True
)
