==> lapse-analysis/5_Target_Variables.ipynb <==
import boto3
import io
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns; sns.set()

import awswrangler as wr

import warnings
warnings.filterwarnings("ignore")

plt.style.use('seaborn-colorblind')

### set-up

session = boto3.Session()
s3 = session.client('s3')

bucket_name = "cdo-ililapse-364524684987-bucket"
file_path = "x266754/lapse/"

### data intake

file_name = "x266754/lapse/lapse_curated_jan_jun.parquet"

%%time
obj = s3.get_object(Bucket = bucket_name, Key = file_name)
df = pd.read_parquet(io.BytesIO(obj['Body'].read())) 

df =df.sort_values(['policy_id', 'pfmc_cur_month'])

df[df['policy_id']=='V2680808'][['policy_id', 'month', 'lapse_ind', 'surrender_ind', 'reinstate_ind']]

### Create 3month ahead status

df['lapse_ind']= df['lapse_ind'].fillna(0)
df['reinstate_ind']= df['reinstate_ind'].fillna(0)
df['surrender_ind']= df['surrender_ind'].fillna(0)

df['cum_lapse_ind'] = df.groupby(['policy_id'])['lapse_ind'].cumsum()
df['cum_reinstate_ind'] = df.groupby(['policy_id'])['reinstate_ind'].cumsum()
df['cum_surrender_ind'] = df.groupby(['policy_id'])['surrender_ind'].cumsum()

df['3mo_ahead_status']= np.where(df['lapse_ind']==1, 'lapse', 
                           (np.where(df['surrender_ind']==1, 'surrender', 
                             (np.where(df['reinstate_ind']==1, 'reinstated',  "current")))))

### Surrender: terminal event, remove rows after surrender

# before
df[df['policy_id']=='V9063643'][['policy_id', 'month', 
                                 'lapse_ind', 
                                 'surrender_ind', 
                                 'reinstate_ind', 
                                 '3mo_ahead_status']]

# remove rows after first surrender occurence 
df['cum_surrender_ind2'] = df.groupby(['policy_id'])['cum_surrender_ind'].cumsum()
df= df[df['cum_surrender_ind2']<2]
del df['cum_surrender_ind2']

# after
df[df['policy_id']=='V9063643'][['policy_id', 'month', 
                                 'lapse_ind', 
                                 'surrender_ind', 
                                 'reinstate_ind', 
                                 '3mo_ahead_status']]

### Lapse: forward fill until reinstated

# before
df[df['policy_id']=='V9051243'][['policy_id', 'month', 
                                 'lapse_ind', 
                                 'surrender_ind', 
                                 'reinstate_ind', 
                                 '3mo_ahead_status']]

df['cum_lapse_reinstate'] = df['cum_lapse_ind'] + df['cum_reinstate_ind']

df['3mo_ahead_status'] = np.where((df['cum_lapse_reinstate']==1), "lapse", df['3mo_ahead_status'])

df.drop(['cum_lapse_ind', 'cum_reinstate_ind', 'cum_surrender_ind'], axis=1, inplace=True)

# after
df[df['policy_id']=='V9051243'][['policy_id', 'month', 
                                 'lapse_ind', 
                                 'surrender_ind', 
                                 'reinstate_ind', 
                                 '3mo_ahead_status']]

### Create current month status

df['current_status'] = df.groupby('agmt_pkge_id')['3mo_ahead_status'].shift(3)
df['current_status']=df['current_status'].fillna('current')

# after
df[df['policy_id']=='V9063643'][['policy_id', 'month', 
                                 'lapse_ind', 
                                 'surrender_ind', 
                                 'reinstate_ind', 
                                 'current_status',
                                 '3mo_ahead_status']]

### sample transition matrix

sample = df[df['pfmc_cur_month']=='2022-03']
tmatrix1 = pd.crosstab(index=sample['current_status'], columns=[sample['3mo_ahead_status']])
display(tmatrix1)

tmatrix2 = pd.crosstab(index=sample['current_status'], columns=[sample['3mo_ahead_status']], normalize='index').round(4)*100
display(tmatrix2)

### export

#!pip install awswrangler --q

%%time
import awswrangler as wr

wr.s3.to_parquet(
    df=df,
    path='s3://cdo-ililapse-364524684987-bucket/x266754/lapse/curated_test.parquet'
)
