==> lapse-analysis/6_Features.ipynb <==
import boto3
import io
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns; sns.set()

import awswrangler as wr

import warnings
warnings.filterwarnings("ignore")

plt.style.use('seaborn-colorblind')

### set-up

session = boto3.Session()
s3 = session.client('s3')

bucket_name = "cdo-ililapse-364524684987-bucket"
file_path = "x266754/lapse/"

### data intake

file_name = "x266754/lapse/curated_test.parquet"

%%time
obj = s3.get_object(Bucket = bucket_name, Key = file_name)
df = pd.read_parquet(io.BytesIO(obj['Body'].read())) 

df[df['policy_id']=='V9063643'][['policy_id', 'month', 
                                 'lapse_ind', 
                                 'surrender_ind', 
                                 'reinstate_ind', 
                                 'current_status',
                                 '3mo_ahead_status']]

### Target Variable

df['3mo_ahead_Lapse'] = (np.where((df['3mo_ahead_status']=='lapse') | ((df['3mo_ahead_status']=='surrender')), 1,0))
df['current_lapse_status'] = (np.where((df['current_status']=='lapse') | ((df['current_status']=='surrender')), 1,0))

### Features

# policy holder
df['holder_age'] = df['attained_age']

# policy age
df['policy_age'] = np.round((df['val_dt'] -df['issue_dt'].astype('datetime64[ns]'))/np.timedelta64(1, 'M'), 0).astype(int)

# issue year
df['issue_year']=  df['issue_dt'].astype('datetime64[ns]').dt.year

# cash value
df['mod_mpt_total'] = df['mpt_tot_act_val'] - df['tot_loan_prncpl']
df['mod_mpt3'] = df['mpt_3'] - df['tot_loan_prncpl']
df['mod_mpt4'] = df['mpt_4'] - df['tot_loan_prncpl']

# payment
df['mod_bld_payment'] = df['modal_prem_bld'] + df['unpd_loan_int_due']
df['mod_sch_payment'] = df['modal_prem_sch'] + df['unpd_loan_int_due']

# cash value - payment
df['cash_accum_prem'] = df['mpt_tot_act_val'] - df['accum_actual_prem_paid']
df['diff_mod_bld_value'] = df['mod_mpt_total'] - df['mod_bld_payment']

# normalized
df['mpt_total_bld_face_amt'] = df['mod_mpt_total']/df['base_face_amt']

# month over month change in cash value
df['mod_mpt_total_lag1'] = df.groupby(['policy_id'])['mod_mpt_total'].shift(1)
df['mod_mpt_total_lag1'] = np.where(df['mod_mpt_total_lag1'].isnull(), df['mod_mpt_total'], df['mod_mpt_total_lag1'])
df['mom_mod_mpt_total']= df['mod_mpt_total'] - df['mod_mpt_total_lag1'] 

# month over month change in payment 
df['accum_actual_prem_paid_lag1'] = df.groupby(['policy_id'])['accum_actual_prem_paid'].shift(1)
df['accum_actual_prem_paid_lag1'] = np.where(df['accum_actual_prem_paid_lag1'].isnull(), df['accum_actual_prem_paid'], df['accum_actual_prem_paid_lag1'])
df['mom_accum_prem_paid']= df['accum_actual_prem_paid'] - df['accum_actual_prem_paid_lag1'] 

# month over month change in prem mode
df['prem_mode_lag1'] = df.groupby(['policy_id'])['prem_mode'].shift(1)
df['prem_mode_lag1'] = np.where(df['prem_mode_lag1'].isnull(), df['prem_mode'], df['prem_mode_lag1'])
df['prem_mode_change_flag']= np.where((df['prem_mode_lag1'] !=df['prem_mode']), 1,0)

# household AUM
df['diff_household_asset_cash_val'] = df['wc_tot_asset_amt']  - df['mod_mpt_total']
df['hh_stock_mix'] = df['wc_asset_mix_stk_amt']/df['wc_tot_asset_amt'] 
df['hh_bond_mix'] = df['wc_asset_mix_bond_amt']/df['wc_tot_asset_amt'] 
df['hh_annuity_mix'] = df['wc_asset_mix_anty_amt']/df['wc_tot_asset_amt'] 

# subject aum
df['diff_aum_cash_val'] = df['subj_aum_amt']  - df['mod_mpt_total']

# other mpt
df['rest_mpt'] = df['mpt_10']+df['mpt_600']+df['mpt_620']+df['mpt_1600']+df['mpt_20881']

# rider flag
df['rdr_face_amt'] = df['rdr1_face_amt']+df['rdr2_face_amt']+df['rdr3_face_amt']
df['rider_flag'] =np.where(df['rdr_face_amt']>0, 1, 0)

# interaction
df['holder_age_cash']=df['holder_age']*df['mod_mpt_total']
df['policy_age_cash']=df['policy_age']/12*df['mod_mpt_total']
df['issue_age_cash']=df['issue_age']*df['mod_mpt_total']

df['diff_policy_issue_issue']=df['policy_age_cash']-df['issue_age_cash']
df['diff_cash_holder_issue']=df['holder_age_cash']-df['issue_age_cash']

# rolling count
df['current'] = df['current_status'].map(lambda x: 1 if x == 'current' else 0)
df['lapse'] = df['current_status'].map(lambda x: 1 if x == 'lapse' else 0)
df['reinstate'] = df['current_status'].map(lambda x: 1 if x == 'reinstated' else 0)

%%time
df['num_current_3m'] = df.groupby('policy_id')['current'].rolling(window=3, min_periods=1).sum().reset_index(level=0, drop=True)
df['num_lapse_3m'] = df.groupby('policy_id')['lapse'].rolling(window=3, min_periods=1).sum().reset_index(level=0, drop=True)
df['num_reinstate_3m'] = df.groupby('policy_id')['reinstate'].rolling(window=3, min_periods=1).sum().reset_index(level=0, drop=True)

#df['num_current_3m'] = df.groupby('policy_id')['current'].transform(lambda x: x.rolling(3).sum())
# df['num_lapse_3m'] = df.groupby('policy_id')['lapse'].transform(lambda x: x.rolling(3).sum())
# df['num_reinstate_3m'] = df.groupby('policy_id')['reinstate'].transform(lambda x: x.rolling(3).sum())

# backward fill
df['num_current_3m']=df['num_current_3m'].bfill()
df['num_reinstate_3m']=df['num_reinstate_3m'].bfill()
df['num_lapse_3m']=df['num_lapse_3m'].bfill()

# number of policies per household and month over month change
count_df = pd.DataFrame(df[(df['household_id']!=888888888)].groupby(['household_id','val_dt'])['agmt_pkge_id'].count()).reset_index()
count_df.columns = ['household_id', 'val_dt', 'num_policies_hh']

%%time
df = df.merge(count_df, how= 'left', 
                      left_on =['household_id','val_dt'],
                      right_on =['household_id','val_dt']
                     )

# month over month change in number of policies per household
df['num_policies_hh_lag1'] = df.groupby(['policy_id'])['num_policies_hh'].shift(1)
df['num_policies_hh_lag1'] = np.where(df['num_policies_hh_lag1'].isnull(), df['num_policies_hh'], df['num_policies_hh_lag1'])
df['mom_hh_policies']= df['num_policies_hh'] - df['num_policies_hh_lag1'] 

# number of service requests last 6 months
df['num_sr_catg'] = df['num_sr_catg'].fillna(0)
df['max_sr_time'] = df['max_sr_time'].fillna(0)
df['time_to_close'] = df['time_to_close'].fillna(0)

df['num_sr_6m'] = df.groupby('policy_id')['num_sr_catg'].rolling(window=6, min_periods=1).sum().reset_index(level=0, drop=True)
df['num_sr_6m']=df['num_sr_6m'].bfill()

# current month - last service request month 
# sr_month_df = pd.DataFrame(df[df['num_sr_catg']>=1].groupby(['policy_id'])['pfmc_cur_month'].last()).reset_index()
# sr_month_df.columns = ['policy_id', 'last_sr_month']

# %%time
# df = df.merge(sr_month_df, how= 'left', 
#                       left_on =['policy_id','pfmc_cur_month'],
#                       right_on =['policy_id','last_sr_month']
#                      )

### buckets 

#age bucket
col         = 'holder_age'  
conditions  = [(df[col] <= 20), (df[col] >20) & (df[col] <= 30),  
               (df[col] >30) & (df[col] <= 35),  (df[col] >35) & (df[col] <= 40), 
               (df[col] >40) & (df[col] <= 45), (df[col] >45) & (df[col] <= 50),  
               (df[col] >50) & (df[col] <= 55), (df[col] >55) & (df[col] <= 60), 
               (df[col] >60) & (df[col] <= 65), (df[col] >65) & (df[col] <= 70), 
               (df[col] >70) & (df[col] <= 75), (df[col] >75) & (df[col] <= 80), 
               (df[col] >80) & (df[col] <= 85), df[col] >85]  
              
choices     = [ '1. <20', '2. (20,30]', '3. (30,35]', '4. (35,40]', '5. (40,45]', '6. (45,50]',
               '7. (50,55]', '8. (55,60]', '9. (60,65]', '10. (65,70]', '11. (70,75]','12. (75,80]', 
               '13. (80,85]','14. >85']
    
df["holder_age_bucket"] = np.select(conditions, choices, default='missing')

#cash value
col         = 'mod_mpt_total'  
conditions  = [(df[col] >=0) & (df[col] <= 10000),
               (df[col] >10000) & (df[col] <= 25000),
               (df[col] >25000) & (df[col] <= 50000),
               (df[col] >50000) & (df[col] <= 100000),  
               (df[col] >100000) & (df[col] <= 250000),  
               (df[col] >250000) & (df[col] <= 500000),  
               (df[col] >500000) & (df[col] <= 750000), 
               (df[col] >750000) & (df[col] <= 1000000),
               (df[col] >1000000) & (df[col] <= 3000000),
               (df[col] >3000000) & (df[col] <= 5000000), 
               (df[col] > 5000000)]  
              
choices     = [ '1. $0k-10k', '2. $10k-25k', '3. $25k-50k','4. $50k-100k', 
               '5. $100k-250k', '6. $250k-500k','7. $500k-750k', '8. $750k-1M', 
               '9. $1M-3M', '10. $3M-5M','11. >$5M']
    
df["cash_val_bucket"] = np.select(conditions, choices, default='missing')

pd.set_option('display.max_columns', None)
df[df['policy_id']=='V9063643'][['policy_id', 'household_id', 
                                 'month', 
                                 'num_policies_hh', 'mom_hh_policies',
                                 'holder_age',  "holder_age_bucket",
                                 'issue_year', 'policy_age',
                                 'mod_mpt_total', 'cash_val_bucket',
                                 'mod_bld_payment', 
                                 'cash_accum_prem', 'diff_mod_bld_value',
                                 'mpt_total_bld_face_amt',  
                                 'accum_actual_prem_paid', 'mom_accum_prem_paid',
                                 'prem_mode', 'prem_mode_change_flag', 
                                 'diff_household_asset_cash_val', 
                                 'hh_stock_mix', 'hh_bond_mix', 'hh_annuity_mix',
                                 'diff_aum_cash_val', 
                                 'rest_mpt',
                                 'rider_flag',
                                 'diff_policy_issue_issue',
                                 'diff_cash_holder_issue',
                                 'num_current_3m',
                                 'num_lapse_3m',
                                 'num_reinstate_3m',
                                 'num_sr_catg',
                                 'num_sr_6m',
                                # 'last_sr_month',
                                # 'month_since_last_sr',
                                 '3mo_ahead_status']]

### Data Filters

# remove decseased  (2 = deceased)
#df= df[df['deceasead_ind']!=2]

# remove policies with zero base face amount 
df = df[df['base_face_amt']!=0]

# remove negative policies with negative policy age
df = df[df['policy_age']>=0]

### export features 

!pip install awswrangler --q

%%time
import awswrangler as wr

wr.s3.to_parquet(
    df=df,
    path='s3://cdo-ililapse-364524684987-bucket/x266754/lapse/curated_features.parquet'
)
